backend/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.13.1
annotated-types==0.7.0
anyio==4.4.0
asyncpg==0.29.0
attrs==23.2.0
Authlib==1.3.1
bcrypt==4.1.3
certifi==2024.6.2
cffi==1.16.0
click==8.1.7
cryptography==42.0.8
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.1.1
fastapi==0.111.0
fastapi-cache2==0.2.1
fastapi-cli==0.0.4
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
multidict==6.0.5
orjson==3.10.4
packaging==24.1
passlib==1.7.4
pendulum==3.0.0
pluggy==1.5.0
psycopg2-binary==2.9.9
pyasn1==0.6.0
pycparser==2.22
pydantic==2.7.4
pydantic-extra-types==2.8.1
pydantic-settings==2.3.3
pydantic_core==2.18.4
Pygments==2.18.0
pytest==8.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
PyYAML==6.0.1
redis==4.6.0
rich==13.7.1
rsa==4.9
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.30
starlette==0.37.2
time-machine==2.14.2
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
uvicorn==0.30.1
uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0
yarl==1.9.4
